Refine cleanup refactors and registration validation
