+++ docker-compose.production.yml
@@ -120,6 +120,9 @@ services:
       AI_DOCUMENT_SERVICE_URL: http://python-ai-docs:8002
       ONLYOFFICE_INTERNAL_URL: http://onlyoffice
       ONLYOFFICE_LARAVEL_INTERNAL_URL: http://laravel:8000
+    # Laravel tetap berjalan via `artisan serve` di port internal 8000.
+    # Caddy menjadi reverse proxy publik, sementara Horizon dan scheduler
+    # tetap berjalan sebagai service terpisah.
     command: ["php", "artisan", "serve", "--host=0.0.0.0", "--port=8000"]
     volumes:
       - laravel_storage:/var/www/html/storage
