+++ python-ai/app/llm_manager.py
@@ -1,10 +1,17 @@
-import json
 import logging
-import warnings
-import requests
-import litellm
-from typing import List, Dict, Generator
+from typing import Dict, Generator, List
+
 from app.env_utils import get_env
+from app.services.llm_streaming import (
+    build_enhanced_messages,
+    compose_enhanced_system_prompt,
+    extract_web_sources,
+    get_chat_models_fallback as _shared_get_chat_models_fallback,
+    get_default_system_prompt_fallback as _shared_get_default_system_prompt_fallback,
+    is_context_too_large as _shared_is_context_too_large,
+    is_rate_limit as _shared_is_rate_limit,
+    stream_with_cascade as _shared_stream_with_cascade,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -41,22 +48,6 @@ except ImportError:
         }
     }
 
-# Suppress verbose litellm output
-litellm.set_verbose = False
-litellm.suppress_debug_info = True
-
-# litellm menggunakan httpx.AsyncClient secara internal untuk streaming.
-# Ketika kita mengkonsumsi stream dalam sync generator, aclose() tidak bisa
-# di-await dengan benar → warning asyncio "Task was destroyed but pending".
-# Ini hanya cosmetic warning, tidak mempengaruhi fungsionalitas.
-warnings.filterwarnings(
-    "ignore",
-    message="coroutine 'AsyncClient.aclose' was never awaited",
-    category=RuntimeWarning,
-)
-# Suppress asyncio logger noise dari masalah yang sama
-logging.getLogger("asyncio").setLevel(logging.CRITICAL)
-
 
 def get_context_for_query(*args, **kwargs):
     from app.services.rag_policy import get_context_for_query as _get_context_for_query
@@ -64,222 +55,38 @@ def get_context_for_query(*args, **kwargs):
     return _get_context_for_query(*args, **kwargs)
 
 
-# ─── Error Classification Helpers ─────────────────────────────────────────────
-
 def _is_context_too_large(error: Exception) -> bool:
-    """Detect 413 / request-body-too-large errors from any provider."""
-    msg = str(error).lower()
-    return any(k in msg for k in [
-        "413", "tokens_limit_reached", "request body too large",
-        "max size", "context_length_exceeded", "too large",
-    ])
+    return _shared_is_context_too_large(error)
+
 
 def _is_rate_limit(error: Exception) -> bool:
-    """Detect 429 / quota-exhausted errors from any provider."""
-    msg = str(error).lower()
-    return any(k in msg for k in [
-        "429", "rate limit", "resource_exhausted", "quota",
-        "too many requests", "503",
-    ])
+    return _shared_is_rate_limit(error)
 
-
-# ─── Model & Prompt Helpers ────────────────────────────────────────────────────
 
 def _get_chat_models_fallback():
-    """Get chat models from config (source of truth)."""
-    if CONFIG_AVAILABLE:
-        models = get_chat_models()
-        if models:
-            return models
-    return []
+    get_chat_models_fn = get_chat_models if CONFIG_AVAILABLE else (lambda: [])
+    return _shared_get_chat_models_fallback(CONFIG_AVAILABLE, get_chat_models_fn)
 
 
 def _get_default_system_prompt_fallback():
-    """Get system prompt from config, env override, or shared default prompt."""
-    if CONFIG_AVAILABLE:
-        try:
-            prompt = get_system_prompt()
-            if prompt:
-                return prompt
-        except Exception as exc:
-            logger.warning("⚠️  Gagal memuat system prompt dari config: %s", exc)
-
     env_prompt = get_env("DEFAULT_SYSTEM_PROMPT", "") or ""
-    if env_prompt:
-        return env_prompt
-
-    return DEFAULT_PROMPTS["system"]["default"]
-
-
-# ─── Gemini Native Streaming ───────────────────────────────────────────────────
-
-def _stream_gemini_native(model_name: str, api_key: str, messages: List[Dict[str, str]]) -> Generator[str, None, None]:
-    """
-    Stream response from Google AI Studio REST API directly.
-
-    Catatan: Gemini systemInstruction punya batas efektif ~8K token.
-    Jika konteks RAG (system message) melebihi 7000 token, konten dipindahkan
-    ke dalam user message pertama agar tidak di-drop diam-diam oleh API.
-    """
-    url = (
-        f"https://generativelanguage.googleapis.com/v1beta/models/"
-        f"{model_name}:streamGenerateContent?alt=sse&key={api_key}"
+    get_system_prompt_fn = get_system_prompt if CONFIG_AVAILABLE else (lambda: "")
+    return _shared_get_default_system_prompt_fallback(
+        CONFIG_AVAILABLE,
+        get_system_prompt_fn,
+        env_prompt,
+        DEFAULT_PROMPTS["system"]["default"],
+        logger,
     )
-
-    # Pisahkan system message dan user/assistant messages
-    system_text = ""
-    user_messages = []
-    for msg in messages:
-        if msg["role"] == "system":
-            system_text = msg["content"]
-        else:
-            role = "user" if msg["role"] == "user" else "model"
-            user_messages.append({"role": role, "parts": [{"text": msg["content"]}]})
-
-    # Cek apakah system message aman untuk systemInstruction (<= 7000 token)
-    GEMINI_SYSTEM_TOKEN_LIMIT = 7000
-    try:
-        import tiktoken as _tiktoken
-        _enc = _tiktoken.get_encoding("cl100k_base")
-        system_tokens = len(_enc.encode(system_text)) if system_text else 0
-    except Exception:
-        system_tokens = len(system_text.split()) if system_text else 0
-
-    contents = []
-    body_system_instruction = None
-
-    if system_text and system_tokens > GEMINI_SYSTEM_TOKEN_LIMIT:
-        # Konteks RAG terlalu besar untuk systemInstruction → injeksi ke user message pertama
-        logger.warning(
-            "⚠️  Gemini: systemInstruction terlalu besar (%d token > %d limit) "
-            "→ konteks RAG dipindahkan ke user message pertama",
-            system_tokens, GEMINI_SYSTEM_TOKEN_LIMIT,
-        )
-        # Semua konten (termasuk teks sistem) masuk ke contents[]
-        if user_messages:
-            # Sisipkan instruksi sistem sebagai prefix pesan user pertama
-            first_user = user_messages[0]
-            first_user["parts"] = [{"text": system_text + "\n\n---\n\n" + first_user["parts"][0]["text"]}]
-            contents = user_messages
-        else:
-            contents = [{"role": "user", "parts": [{"text": system_text}]}]
-    else:
-        contents = user_messages
-        if system_text:
-            body_system_instruction = {"parts": [{"text": system_text}]}
-
-    body = {"contents": contents}
-    if body_system_instruction:
-        body["systemInstruction"] = body_system_instruction
-
-    response = requests.post(url, json=body, stream=True, timeout=30)
-    response.raise_for_status()
-
-    for line in response.iter_lines():
-        if line:
-            decoded = line.decode("utf-8")
-            if decoded.startswith("data: "):
-                try:
-                    data = json.loads(decoded[6:])
-                    text = (
-                        data.get("candidates", [{}])[0]
-                        .get("content", {})
-                        .get("parts", [{}])[0]
-                        .get("text", "")
-                    )
-                    if text:
-                        yield text
-                except (json.JSONDecodeError, IndexError, KeyError):
-                    continue
-
-
-# ─── Core LLM Streaming with Cascade Fallback ─────────────────────────────────
-
-def _run_model(model: dict, messages: List[Dict[str, str]]) -> Generator:
-    """Create a streaming generator for the given model config."""
-    api_key = get_env(model["api_key_env"])
-    if not api_key:
-        raise ValueError(f"API key env '{model['api_key_env']}' tidak ditemukan")
-
-    if model["provider"] == "gemini_native":
-        return _stream_gemini_native(model["model_name"], api_key, messages)
-
-    kwargs = {
-        "model": model["model_name"],
-        "messages": messages,
-        "api_key": api_key,
-        "stream": True,
-        "timeout": 30,
-        "num_retries": 0,
-    }
-    if "base_url" in model:
-        kwargs["api_base"] = model["base_url"]
-    return litellm.completion(**kwargs)
 
 
 def _stream_with_cascade(
     messages: List[Dict[str, str]],
-    sources: List[Dict] = None,
+    sources: List[Dict] | None = None,
 ) -> Generator[str, None, None]:
-    """
-    Core cascade engine: iterate model list, yield tokens, handle fallback.
-
-    Cascade rules:
-    - 413 / context too large  → skip to next model (larger context window)
-    - 429 / rate limit         → skip to next model
-    - Other errors             → skip to next model
-    All errors are logged clearly so the terminal is readable.
-    """
     model_list = _get_chat_models_fallback()
-    total = len(model_list)
+    yield from _shared_stream_with_cascade(messages, model_list=model_list, sources=sources, logger=logger)
 
-    for idx, model in enumerate(model_list, start=1):
-        label = model["label"]
-        try:
-            gen = _run_model(model, messages)
-            logger.info("🤖 [%d/%d] Menggunakan: %s", idx, total, label)
-            yield f"[MODEL:{label}]\n"
-
-            if model["provider"] == "gemini_native":
-                for chunk in gen:
-                    yield chunk
-            else:
-                for chunk in gen:
-                    content = chunk.choices[0].delta.content
-                    if content:
-                        yield content
-
-            # Append sources kalau ada
-            if sources:
-                sources_json = json.dumps(sources, ensure_ascii=False, separators=(',', ':'))
-                yield f"\n\n[SOURCES:{sources_json}]"
-
-            logger.info("✅ Respons selesai dari: %s", label)
-            return
-
-        except Exception as e:
-            if _is_context_too_large(e):
-                logger.warning(
-                    "⚠️  [%d/%d] %s → konteks terlalu besar (413), cascade ke model berikutnya...",
-                    idx, total, label,
-                )
-            elif _is_rate_limit(e):
-                logger.warning(
-                    "⚠️  [%d/%d] %s → rate limit (429), cascade ke model berikutnya...",
-                    idx, total, label,
-                )
-            else:
-                logger.warning(
-                    "⚠️  [%d/%d] %s → error: %s",
-                    idx, total, label, str(e)[:120],
-                )
-            continue
-
-    logger.error("❌ Semua %d model gagal. Tidak ada respons yang bisa dikirim.", total)
-    yield "❌ Maaf, semua layanan AI sedang tidak tersedia. Silakan coba lagi nanti."
-
-
-# ─── Public API ───────────────────────────────────────────────────────────────
 
 def get_llm_stream(
     messages: List[Dict[str, str]],
@@ -305,7 +112,6 @@ def get_llm_stream(
 
     default_system_prompt = _get_default_system_prompt_fallback()
 
-    # Cari konteks dari web search (jika diaktifkan)
     search_context = ""
     web_sources: list = []
     if query:
@@ -318,32 +124,16 @@ def get_llm_stream(
                 explicit_web_request=explicit_web_request,
             )
             search_context = context_data.get("search_context", "")
-
-            # Ekstrak web search results sebagai sources (untuk ditampilkan sebagai link)
-            raw_results = context_data.get("search_results", [])
-            if raw_results:
-                for r in raw_results:
-                    url   = (r.get("url") or "").strip()
-                    title = (r.get("title") or "").strip()
-                    snippet = (r.get("snippet") or r.get("description") or "").strip()
-                    if url:  # hanya sertakan jika ada URL valid
-                        web_sources.append({
-                            "type":     "web",
-                            "title":    title or url,
-                            "url":      url,
-                            "snippet":  snippet[:160] if snippet else "",
-                        })
+            web_sources = extract_web_sources(context_data)
         except Exception as e:
             logger.warning("⚠️  Web search/RAG context gagal: %s", e)
 
-    # Bangun system prompt
     if search_context:
-        assertive_instruction = ""
         if CONFIG_AVAILABLE:
             try:
                 assertive_instruction = get_assertive_instruction()
             except Exception:
-                pass
+                assertive_instruction = ""
         else:
             assertive_instruction = (
                 "\n\nInstruksi tambahan:\n"
@@ -351,22 +141,17 @@ def get_llm_stream(
                 "- Jika sumber web tersedia, utamakan data faktual dari sumber tersebut.\n"
                 "- Jawab secara ringkas, jelas, dan hindari istilah teknis internal sistem."
             )
-
-        base = system_prompt_base if system_prompt_base else default_system_prompt
-        enhanced_system = f"{search_context.rstrip()}\n\n{base}\n\n{assertive_instruction.strip()}".strip()
     else:
-        enhanced_system = system_prompt_base if system_prompt_base else default_system_prompt
+        assertive_instruction = ""
 
-    # Susun pesan final
-    enhanced_messages = []
-    has_system_message = any(msg["role"] == "system" for msg in messages)
-    if not has_system_message:
-        enhanced_messages.append({"role": "system", "content": enhanced_system})
-    for msg in messages:
-        if msg["role"] == "system":
-            enhanced_messages.append({"role": "system", "content": enhanced_system})
-        else:
-            enhanced_messages.append(msg)
+    enhanced_system = compose_enhanced_system_prompt(
+        search_context=search_context,
+        system_prompt_base=system_prompt_base,
+        default_system_prompt=default_system_prompt,
+        assertive_instruction=assertive_instruction,
+    )
+
+    enhanced_messages = build_enhanced_messages(messages, enhanced_system)
 
     yield from _stream_with_cascade(enhanced_messages, sources=web_sources or None)
