+++ python-ai/app/services/rag_ingest.py
@@ -1,16 +1,9 @@
 import os
-import csv
 import hashlib
 import logging
 import time
-from pathlib import Path
-from typing import List, Tuple
 
 from langchain_chroma import Chroma
-from langchain_core.documents import Document
-import pdfplumber
-from docx import Document as DocxDocument
-from openpyxl import load_workbook
 
 from app.services.rag_config import (
     CHROMA_PATH,
@@ -25,144 +18,10 @@ from app.services.rag_config import (
     EMBEDDING_MODELS,
 )
 from app.services.rag_embeddings import count_tokens, get_embeddings_with_fallback
+from app.services.document_loaders import load_documents_lightweight as _load_documents_lightweight
 from app.services.lightweight_text_splitter import LightweightRecursiveTextSplitter
 
 logger = logging.getLogger(__name__)
-
-
-def _clean_text(value) -> str:
-    return "" if value is None else str(value).strip()
-
-
-def _table_to_text(rows: list[list[object]]) -> str:
-    lines: list[str] = []
-    for row in rows:
-        cleaned = [_clean_text(cell) for cell in row]
-        if any(cleaned):
-            lines.append(" | ".join(cleaned))
-    return "\n".join(lines)
-
-
-def _load_pdf_documents(file_path: str, filename: str) -> list[Document]:
-    docs: list[Document] = []
-
-    with pdfplumber.open(file_path) as pdf:
-        for page_number, page in enumerate(pdf.pages, start=1):
-            parts: list[str] = []
-            text = (page.extract_text() or "").strip()
-            if text:
-                parts.append(text)
-
-            for table in page.extract_tables() or []:
-                table_text = _table_to_text(table)
-                if table_text:
-                    parts.append(table_text)
-
-            content = "\n\n".join(parts).strip()
-            if content:
-                docs.append(
-                    Document(
-                        page_content=content,
-                        metadata={"source": filename, "page": page_number},
-                    )
-                )
-
-    return docs
-
-
-def _load_docx_documents(file_path: str, filename: str) -> list[Document]:
-    document = DocxDocument(file_path)
-    parts: list[str] = []
-
-    for paragraph in document.paragraphs:
-        text = paragraph.text.strip()
-        if text:
-            parts.append(text)
-
-    for table in document.tables:
-        table_text = _table_to_text(
-            [[cell.text for cell in row.cells] for row in table.rows]
-        )
-        if table_text:
-            parts.append(table_text)
-
-    content = "\n\n".join(parts).strip()
-    return [Document(page_content=content, metadata={"source": filename})] if content else []
-
-
-def _load_xlsx_documents(file_path: str, filename: str) -> list[Document]:
-    workbook = load_workbook(file_path, data_only=True, read_only=True)
-    docs: list[Document] = []
-
-    try:
-        for sheet in workbook.worksheets:
-            rows = [
-                [cell for cell in row]
-                for row in sheet.iter_rows(values_only=True)
-            ]
-            table_text = _table_to_text(rows)
-            if table_text:
-                docs.append(
-                    Document(
-                        page_content=f"{sheet.title}\n\n{table_text}",
-                        metadata={"source": filename, "sheet": sheet.title},
-                    )
-                )
-    finally:
-        workbook.close()
-
-    return docs
-
-
-def _load_csv_documents(file_path: str, filename: str) -> list[Document]:
-    rows: list[list[str]] = []
-
-    for encoding in ("utf-8-sig", "utf-8", "latin-1"):
-        try:
-            with open(file_path, newline="", encoding=encoding) as handle:
-                sample = handle.read(4096)
-                handle.seek(0)
-                try:
-                    dialect = csv.Sniffer().sniff(sample)
-                except csv.Error:
-                    dialect = csv.excel
-                rows = [row for row in csv.reader(handle, dialect)]
-            break
-        except UnicodeDecodeError:
-            continue
-
-    content = _table_to_text(rows)
-    return [Document(page_content=content, metadata={"source": filename})] if content else []
-
-
-def _load_text_documents(file_path: str, filename: str) -> list[Document]:
-    for encoding in ("utf-8-sig", "utf-8", "latin-1"):
-        try:
-            content = Path(file_path).read_text(encoding=encoding).strip()
-            if content:
-                return [Document(page_content=content, metadata={"source": filename})]
-            return []
-        except UnicodeDecodeError:
-            continue
-
-    return []
-
-
-def _load_documents_lightweight(file_path: str, filename: str) -> list[Document]:
-    suffix = Path(filename).suffix.lower()
-
-    if suffix == ".pdf":
-        return _load_pdf_documents(file_path, filename)
-    if suffix == ".docx":
-        return _load_docx_documents(file_path, filename)
-    if suffix == ".xlsx":
-        return _load_xlsx_documents(file_path, filename)
-    if suffix == ".csv":
-        return _load_csv_documents(file_path, filename)
-    if suffix in {".txt", ".md", ".markdown"}:
-        return _load_text_documents(file_path, filename)
-
-    raise ValueError(f"Unsupported document type for lightweight ingest: {suffix or 'unknown'}")
 
 
 def process_document(file_path: str, filename: str, user_id: str = "unknown"):
